Refactor Android Room DAOs
