--- apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/database/DatabaseDaos.kt
@@ -1,1854 +1,0 @@
-package com.flashcardsopensourceapp.data.local.database
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import androidx.room.Transaction
-import androidx.room.Update
-import com.flashcardsopensourceapp.data.local.model.EffortLevel
-import kotlinx.coroutines.flow.Flow
-
-internal const val activeReviewRecentPriorityWindowMillis: Long = 60L * 60L * 1_000L
-
-data class ReviewEffortCountRow(
-    val effortLevel: com.flashcardsopensourceapp.data.local.model.EffortLevel,
-    val totalCount: Int
-)
-
-data class ReviewTagCountRow(
-    val tag: String,
-    val totalCount: Int
-)
-
-@Dao
-interface AppLocalSettingsDao {
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertSettings(settings: AppLocalSettingsEntity)
-
-    @Update
-    suspend fun updateSettings(settings: AppLocalSettingsEntity)
-
-    @Query("SELECT * FROM app_local_settings WHERE settingsId = 1 LIMIT 1")
-    fun observeSettings(): Flow<AppLocalSettingsEntity?>
-
-    @Query("SELECT * FROM app_local_settings WHERE settingsId = 1 LIMIT 1")
-    suspend fun loadSettings(): AppLocalSettingsEntity?
-
-    @Query("DELETE FROM app_local_settings")
-    suspend fun deleteAllSettings()
-}
-
-@Dao
-interface WorkspaceDao {
-    @Insert(onConflict = OnConflictStrategy.ABORT)
-    suspend fun insertWorkspace(workspace: WorkspaceEntity)
-
-    @Query("SELECT COUNT(*) FROM workspaces")
-    suspend fun countWorkspaces(): Int
-
-    @Query("SELECT * FROM workspaces ORDER BY createdAtMillis DESC, workspaceId DESC")
-    fun observeWorkspaces(): Flow<List<WorkspaceEntity>>
-
-    @Query("SELECT * FROM workspaces ORDER BY createdAtMillis DESC, workspaceId DESC")
-    suspend fun loadWorkspaces(): List<WorkspaceEntity>
-
-    @Query("SELECT * FROM workspaces WHERE workspaceId = :workspaceId LIMIT 1")
-    fun observeWorkspaceById(workspaceId: String): Flow<WorkspaceEntity?>
-
-    @Query("SELECT * FROM workspaces WHERE workspaceId = :workspaceId LIMIT 1")
-    suspend fun loadWorkspaceById(workspaceId: String): WorkspaceEntity?
-
-    @Query("SELECT * FROM workspaces ORDER BY createdAtMillis ASC, workspaceId ASC LIMIT 1")
-    fun observeAnyWorkspace(): Flow<WorkspaceEntity?>
-
-    @Query("SELECT * FROM workspaces ORDER BY createdAtMillis ASC, workspaceId ASC LIMIT 1")
-    suspend fun loadAnyWorkspace(): WorkspaceEntity?
-
-    @Update
-    suspend fun updateWorkspace(workspace: WorkspaceEntity)
-
-    @Query("DELETE FROM workspaces")
-    suspend fun deleteAllWorkspaces()
-
-    @Query("DELETE FROM workspaces WHERE workspaceId = :workspaceId")
-    suspend fun deleteWorkspace(workspaceId: String)
-}
-
-@Dao
-interface WorkspaceSchedulerSettingsDao {
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertWorkspaceSchedulerSettings(settings: WorkspaceSchedulerSettingsEntity)
-
-    @Update
-    suspend fun updateWorkspaceSchedulerSettings(settings: WorkspaceSchedulerSettingsEntity)
-
-    @Query("SELECT * FROM workspace_scheduler_settings WHERE workspaceId = :workspaceId LIMIT 1")
-    fun observeWorkspaceSchedulerSettings(workspaceId: String): Flow<WorkspaceSchedulerSettingsEntity?>
-
-    @Query("SELECT * FROM workspace_scheduler_settings WHERE workspaceId = :workspaceId LIMIT 1")
-    suspend fun loadWorkspaceSchedulerSettings(workspaceId: String): WorkspaceSchedulerSettingsEntity?
-
-    @Query("UPDATE workspace_scheduler_settings SET workspaceId = :newWorkspaceId WHERE workspaceId = :oldWorkspaceId")
-    suspend fun reassignWorkspace(oldWorkspaceId: String, newWorkspaceId: String)
-}
-
-@Dao
-interface DeckDao {
-    @Insert(onConflict = OnConflictStrategy.ABORT)
-    suspend fun insertDeck(deck: DeckEntity)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertDecks(decks: List<DeckEntity>)
-
-    @Update
-    suspend fun updateDeck(deck: DeckEntity)
-
-    @Query("DELETE FROM decks WHERE deckId = :deckId")
-    suspend fun deleteDeck(deckId: String)
-
-    @Query("SELECT * FROM decks ORDER BY createdAtMillis DESC, deckId DESC")
-    fun observeDecks(): Flow<List<DeckEntity>>
-
-    @Query("SELECT * FROM decks WHERE deckId = :deckId LIMIT 1")
-    fun observeDeck(deckId: String): Flow<DeckEntity?>
-
-    @Query("SELECT * FROM decks WHERE deckId = :deckId LIMIT 1")
-    suspend fun loadDeck(deckId: String): DeckEntity?
-
-    @Query("SELECT * FROM decks WHERE workspaceId = :workspaceId ORDER BY createdAtMillis ASC, deckId ASC")
-    suspend fun loadDecks(workspaceId: String): List<DeckEntity>
-
-    @Query("SELECT COUNT(*) FROM decks")
-    fun observeDeckCount(): Flow<Int>
-
-    @Query("SELECT COUNT(*) FROM decks WHERE deletedAtMillis IS NULL")
-    suspend fun countDecks(): Int
-
-    @Query("DELETE FROM decks")
-    suspend fun deleteAllDecks()
-
-    @Query("UPDATE decks SET workspaceId = :newWorkspaceId WHERE workspaceId = :oldWorkspaceId")
-    suspend fun reassignWorkspace(oldWorkspaceId: String, newWorkspaceId: String)
-}
-
-@Dao
-interface CardDao {
-    @Transaction
-    @Query("SELECT * FROM cards ORDER BY updatedAtMillis DESC, createdAtMillis DESC")
-    fun observeCardsWithRelations(): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query("SELECT * FROM cards WHERE cardId = :cardId LIMIT 1")
-    fun observeCardWithRelations(cardId: String): Flow<CardWithRelations?>
-
-    @Transaction
-    @Query("SELECT * FROM cards WHERE cardId = :cardId AND workspaceId = :workspaceId LIMIT 1")
-    fun observeCardWithRelationsByWorkspace(cardId: String, workspaceId: String): Flow<CardWithRelations?>
-
-    @Transaction
-    @Query("SELECT * FROM cards ORDER BY createdAtMillis ASC")
-    fun observeReviewCards(): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeRecentlyReviewedDueReviewQueue(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-            AND effortLevel IN (:effortLevels)
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeRecentlyReviewedDueReviewQueueByEffortLevels(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeRecentlyReviewedDueReviewQueueByAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeRecentlyReviewedDueReviewQueueByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeOtherDueReviewQueue(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-            AND effortLevel IN (:effortLevels)
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeOtherDueReviewQueueByEffortLevels(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeOtherDueReviewQueueByAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeOtherDueReviewQueueByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeNewReviewQueue(
-        workspaceId: String,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-            AND effortLevel IN (:effortLevels)
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeNewReviewQueueByEffortLevels(
-        workspaceId: String,
-        effortLevels: List<EffortLevel>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeNewReviewQueueByAnyTags(
-        workspaceId: String,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeNewReviewQueueByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        effortLevels: List<EffortLevel>,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-        FROM (
-            SELECT 0 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND fsrsLastReviewedAtMillis IS NOT NULL
-                    AND fsrsLastReviewedAtMillis >= :cutoffMillis
-                    AND fsrsLastReviewedAtMillis <= :nowMillis
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 1 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND (
-                        fsrsLastReviewedAtMillis IS NULL
-                        OR fsrsLastReviewedAtMillis < :cutoffMillis
-                        OR fsrsLastReviewedAtMillis > :nowMillis
-                    )
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 2 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NULL
-                ORDER BY createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-        )
-        ORDER BY activeQueueBucket ASC, dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeBucketedActiveReviewQueue(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-        FROM (
-            SELECT 0 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND fsrsLastReviewedAtMillis IS NOT NULL
-                    AND fsrsLastReviewedAtMillis >= :cutoffMillis
-                    AND fsrsLastReviewedAtMillis <= :nowMillis
-                    AND effortLevel IN (:effortLevels)
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 1 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND (
-                        fsrsLastReviewedAtMillis IS NULL
-                        OR fsrsLastReviewedAtMillis < :cutoffMillis
-                        OR fsrsLastReviewedAtMillis > :nowMillis
-                    )
-                    AND effortLevel IN (:effortLevels)
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 2 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NULL
-                    AND effortLevel IN (:effortLevels)
-                ORDER BY createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-        )
-        ORDER BY activeQueueBucket ASC, dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeBucketedActiveReviewQueueByEffortLevels(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-        FROM (
-            SELECT 0 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND fsrsLastReviewedAtMillis IS NOT NULL
-                    AND fsrsLastReviewedAtMillis >= :cutoffMillis
-                    AND fsrsLastReviewedAtMillis <= :nowMillis
-                    AND EXISTS (
-                        SELECT 1
-                        FROM card_tags
-                        INNER JOIN tags ON tags.tagId = card_tags.tagId
-                        WHERE card_tags.cardId = cards.cardId
-                            AND tags.workspaceId = cards.workspaceId
-                            AND tags.name IN (:tagNames)
-                    )
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 1 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND (
-                        fsrsLastReviewedAtMillis IS NULL
-                        OR fsrsLastReviewedAtMillis < :cutoffMillis
-                        OR fsrsLastReviewedAtMillis > :nowMillis
-                    )
-                    AND EXISTS (
-                        SELECT 1
-                        FROM card_tags
-                        INNER JOIN tags ON tags.tagId = card_tags.tagId
-                        WHERE card_tags.cardId = cards.cardId
-                            AND tags.workspaceId = cards.workspaceId
-                            AND tags.name IN (:tagNames)
-                    )
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 2 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NULL
-                    AND EXISTS (
-                        SELECT 1
-                        FROM card_tags
-                        INNER JOIN tags ON tags.tagId = card_tags.tagId
-                        WHERE card_tags.cardId = cards.cardId
-                            AND tags.workspaceId = cards.workspaceId
-                            AND tags.name IN (:tagNames)
-                    )
-                ORDER BY createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-        )
-        ORDER BY activeQueueBucket ASC, dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeBucketedActiveReviewQueueByAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query(
-        """
-        SELECT cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-        FROM (
-            SELECT 0 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND fsrsLastReviewedAtMillis IS NOT NULL
-                    AND fsrsLastReviewedAtMillis >= :cutoffMillis
-                    AND fsrsLastReviewedAtMillis <= :nowMillis
-                    AND effortLevel IN (:effortLevels)
-                    AND EXISTS (
-                        SELECT 1
-                        FROM card_tags
-                        INNER JOIN tags ON tags.tagId = card_tags.tagId
-                        WHERE card_tags.cardId = cards.cardId
-                            AND tags.workspaceId = cards.workspaceId
-                            AND tags.name IN (:tagNames)
-                    )
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 1 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NOT NULL
-                    AND dueAtMillis <= :nowMillis
-                    AND (
-                        fsrsLastReviewedAtMillis IS NULL
-                        OR fsrsLastReviewedAtMillis < :cutoffMillis
-                        OR fsrsLastReviewedAtMillis > :nowMillis
-                    )
-                    AND effortLevel IN (:effortLevels)
-                    AND EXISTS (
-                        SELECT 1
-                        FROM card_tags
-                        INNER JOIN tags ON tags.tagId = card_tags.tagId
-                        WHERE card_tags.cardId = cards.cardId
-                            AND tags.workspaceId = cards.workspaceId
-                            AND tags.name IN (:tagNames)
-                    )
-                ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-            UNION ALL
-            SELECT 2 AS activeQueueBucket, cardId, workspaceId, frontText, backText, effortLevel, dueAtMillis, createdAtMillis, updatedAtMillis, reps, lapses, fsrsCardState, fsrsStepIndex, fsrsStability, fsrsDifficulty, fsrsLastReviewedAtMillis, fsrsScheduledDays, deletedAtMillis
-            FROM (
-                SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-                WHERE workspaceId = :workspaceId
-                    AND deletedAtMillis IS NULL
-                    AND dueAtMillis IS NULL
-                    AND effortLevel IN (:effortLevels)
-                    AND EXISTS (
-                        SELECT 1
-                        FROM card_tags
-                        INNER JOIN tags ON tags.tagId = card_tags.tagId
-                        WHERE card_tags.cardId = cards.cardId
-                            AND tags.workspaceId = cards.workspaceId
-                            AND tags.name IN (:tagNames)
-                    )
-                ORDER BY createdAtMillis DESC, cardId ASC
-                LIMIT :limit
-            )
-        )
-        ORDER BY activeQueueBucket ASC, dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT :limit
-        """
-    )
-    fun observeBucketedActiveReviewQueueByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        tagNames: List<String>,
-        limit: Int
-    ): Flow<List<CardWithRelations>>
-
-    @Transaction
-    @Query("SELECT * FROM cards WHERE workspaceId = :workspaceId AND cardId IN (:cardIds)")
-    fun observeCardsWithRelationsByWorkspaceAndIds(
-        workspaceId: String,
-        cardIds: List<String>
-    ): Flow<List<CardWithRelations>>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-        """
-    )
-    fun observeReviewTotalCount(workspaceId: String): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND effortLevel IN (:effortLevels)
-        """
-    )
-    fun observeReviewTotalCountByEffortLevels(
-        workspaceId: String,
-        effortLevels: List<com.flashcardsopensourceapp.data.local.model.EffortLevel>
-    ): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        """
-    )
-    fun observeReviewTotalCountByAnyTags(
-        workspaceId: String,
-        tagNames: List<String>
-    ): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        """
-    )
-    fun observeReviewTotalCountByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        effortLevels: List<com.flashcardsopensourceapp.data.local.model.EffortLevel>,
-        tagNames: List<String>
-    ): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-        """
-    )
-    fun observeReviewDueCount(workspaceId: String, nowMillis: Long): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-            AND effortLevel IN (:effortLevels)
-        """
-    )
-    fun observeReviewDueCountByEffortLevels(
-        workspaceId: String,
-        nowMillis: Long,
-        effortLevels: List<com.flashcardsopensourceapp.data.local.model.EffortLevel>
-    ): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        """
-    )
-    fun observeReviewDueCountByAnyTags(
-        workspaceId: String,
-        nowMillis: Long,
-        tagNames: List<String>
-    ): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        """
-    )
-    fun observeReviewDueCountByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        nowMillis: Long,
-        effortLevels: List<com.flashcardsopensourceapp.data.local.model.EffortLevel>,
-        tagNames: List<String>
-    ): Flow<Int>
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-        """
-    )
-    suspend fun countReviewDueCards(workspaceId: String, nowMillis: Long): Int
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-            AND effortLevel IN (:effortLevels)
-        """
-    )
-    suspend fun countReviewDueCardsByEffortLevels(
-        workspaceId: String,
-        nowMillis: Long,
-        effortLevels: List<com.flashcardsopensourceapp.data.local.model.EffortLevel>
-    ): Int
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        """
-    )
-    suspend fun countReviewDueCardsByAnyTags(
-        workspaceId: String,
-        nowMillis: Long,
-        tagNames: List<String>
-    ): Int
-
-    @Query(
-        """
-        SELECT COUNT(*) FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        """
-    )
-    suspend fun countReviewDueCardsByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        nowMillis: Long,
-        effortLevels: List<com.flashcardsopensourceapp.data.local.model.EffortLevel>,
-        tagNames: List<String>
-    ): Int
-
-    @Query(
-        """
-        SELECT effortLevel, COUNT(*) AS totalCount
-        FROM cards
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND (dueAtMillis IS NULL OR dueAtMillis <= :nowMillis)
-        GROUP BY effortLevel
-        """
-    )
-    fun observeReviewEffortDueCounts(workspaceId: String, nowMillis: Long): Flow<List<ReviewEffortCountRow>>
-
-    @Query(
-        """
-        SELECT tags.name AS tag, COUNT(DISTINCT cards.cardId) AS totalCount
-        FROM cards
-        INNER JOIN card_tags ON card_tags.cardId = cards.cardId
-        INNER JOIN tags ON tags.tagId = card_tags.tagId
-        WHERE cards.workspaceId = :workspaceId
-            AND tags.workspaceId = cards.workspaceId
-            AND cards.deletedAtMillis IS NULL
-            AND (cards.dueAtMillis IS NULL OR cards.dueAtMillis <= :nowMillis)
-        GROUP BY tags.name
-        """
-    )
-    fun observeReviewTagDueCounts(workspaceId: String, nowMillis: Long): Flow<List<ReviewTagCountRow>>
-
-    @Insert(onConflict = OnConflictStrategy.ABORT)
-    suspend fun insertCard(card: CardEntity)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertCards(cards: List<CardEntity>)
-
-    @Update
-    suspend fun updateCard(card: CardEntity)
-
-    @Query("DELETE FROM cards WHERE cardId = :cardId")
-    suspend fun deleteCard(cardId: String)
-
-    @Query("SELECT * FROM cards WHERE cardId = :cardId LIMIT 1")
-    suspend fun loadCard(cardId: String): CardEntity?
-
-    @Query("SELECT * FROM cards WHERE workspaceId = :workspaceId ORDER BY createdAtMillis ASC, cardId ASC")
-    suspend fun loadCards(workspaceId: String): List<CardEntity>
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopRecentlyReviewedDueReviewCard(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-            AND effortLevel IN (:effortLevels)
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopRecentlyReviewedDueReviewCardByEffortLevels(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopRecentlyReviewedDueReviewCardByAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        tagNames: List<String>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND fsrsLastReviewedAtMillis IS NOT NULL
-            AND fsrsLastReviewedAtMillis >= :cutoffMillis
-            AND fsrsLastReviewedAtMillis <= :nowMillis
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopRecentlyReviewedDueReviewCardByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        tagNames: List<String>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopOtherDueReviewCard(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-            AND effortLevel IN (:effortLevels)
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopOtherDueReviewCardByEffortLevels(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopOtherDueReviewCardByAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        tagNames: List<String>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NOT NULL
-            AND dueAtMillis <= :nowMillis
-            AND (
-                fsrsLastReviewedAtMillis IS NULL
-                OR fsrsLastReviewedAtMillis < :cutoffMillis
-                OR fsrsLastReviewedAtMillis > :nowMillis
-            )
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY dueAtMillis ASC, createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopOtherDueReviewCardByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        cutoffMillis: Long,
-        nowMillis: Long,
-        effortLevels: List<EffortLevel>,
-        tagNames: List<String>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopNewReviewCard(workspaceId: String): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-            AND effortLevel IN (:effortLevels)
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopNewReviewCardByEffortLevels(
-        workspaceId: String,
-        effortLevels: List<EffortLevel>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopNewReviewCardByAnyTags(
-        workspaceId: String,
-        tagNames: List<String>
-    ): CardEntity?
-
-    @Query(
-        """
-        SELECT * FROM cards INDEXED BY index_cards_workspaceId_dueAtMillis_createdAtMillis_cardId
-        WHERE workspaceId = :workspaceId
-            AND deletedAtMillis IS NULL
-            AND dueAtMillis IS NULL
-            AND effortLevel IN (:effortLevels)
-            AND EXISTS (
-                SELECT 1
-                FROM card_tags
-                INNER JOIN tags ON tags.tagId = card_tags.tagId
-                WHERE card_tags.cardId = cards.cardId
-                    AND tags.workspaceId = cards.workspaceId
-                    AND tags.name IN (:tagNames)
-            )
-        ORDER BY createdAtMillis DESC, cardId ASC
-        LIMIT 1
-        """
-    )
-    suspend fun loadTopNewReviewCardByEffortLevelsAndAnyTags(
-        workspaceId: String,
-        effortLevels: List<EffortLevel>,
-        tagNames: List<String>
-    ): CardEntity?
-
-    @Query("SELECT COUNT(*) FROM cards")
-    fun observeCardCount(): Flow<Int>
-
-    @Query("SELECT COUNT(*) FROM cards WHERE deletedAtMillis IS NULL")
-    suspend fun countActiveCards(): Int
-
-    @Query(
-        """
-        SELECT cardId, workspaceId, dueAtMillis
-        FROM cards
-        WHERE deletedAtMillis IS NULL
-        ORDER BY workspaceId ASC, dueAtMillis ASC, cardId ASC
-        """
-    )
-    fun observeProgressReviewScheduleCardDueDates(): Flow<List<ProgressReviewScheduleCardDueEntity>>
-
-    @Query("DELETE FROM cards")
-    suspend fun deleteAllCards()
-
-    @Query("UPDATE cards SET workspaceId = :newWorkspaceId WHERE workspaceId = :oldWorkspaceId")
-    suspend fun reassignWorkspace(oldWorkspaceId: String, newWorkspaceId: String)
-}
-
-suspend fun loadTopActiveReviewCard(
-    cardDao: CardDao,
-    workspaceId: String,
-    nowMillis: Long,
-    effortLevels: List<EffortLevel>,
-    tagNames: List<String>
-): CardEntity? {
-    val cutoffMillis = nowMillis - activeReviewRecentPriorityWindowMillis
-    return loadTopRecentlyReviewedDueReviewCardForPredicate(
-        cardDao = cardDao,
-        workspaceId = workspaceId,
-        cutoffMillis = cutoffMillis,
-        nowMillis = nowMillis,
-        effortLevels = effortLevels,
-        tagNames = tagNames
-    ) ?: loadTopOtherDueReviewCardForPredicate(
-        cardDao = cardDao,
-        workspaceId = workspaceId,
-        cutoffMillis = cutoffMillis,
-        nowMillis = nowMillis,
-        effortLevels = effortLevels,
-        tagNames = tagNames
-    ) ?: loadTopNewReviewCardForPredicate(
-        cardDao = cardDao,
-        workspaceId = workspaceId,
-        effortLevels = effortLevels,
-        tagNames = tagNames
-    )
-}
-
-private suspend fun loadTopRecentlyReviewedDueReviewCardForPredicate(
-    cardDao: CardDao,
-    workspaceId: String,
-    cutoffMillis: Long,
-    nowMillis: Long,
-    effortLevels: List<EffortLevel>,
-    tagNames: List<String>
-): CardEntity? {
-    return when {
-        effortLevels.isEmpty() && tagNames.isEmpty() -> cardDao.loadTopRecentlyReviewedDueReviewCard(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis
-        )
-
-        effortLevels.isNotEmpty() && tagNames.isEmpty() -> cardDao.loadTopRecentlyReviewedDueReviewCardByEffortLevels(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis,
-            effortLevels = effortLevels
-        )
-
-        effortLevels.isEmpty() -> cardDao.loadTopRecentlyReviewedDueReviewCardByAnyTags(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis,
-            tagNames = tagNames
-        )
-
-        else -> cardDao.loadTopRecentlyReviewedDueReviewCardByEffortLevelsAndAnyTags(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis,
-            effortLevels = effortLevels,
-            tagNames = tagNames
-        )
-    }
-}
-
-private suspend fun loadTopOtherDueReviewCardForPredicate(
-    cardDao: CardDao,
-    workspaceId: String,
-    cutoffMillis: Long,
-    nowMillis: Long,
-    effortLevels: List<EffortLevel>,
-    tagNames: List<String>
-): CardEntity? {
-    return when {
-        effortLevels.isEmpty() && tagNames.isEmpty() -> cardDao.loadTopOtherDueReviewCard(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis
-        )
-
-        effortLevels.isNotEmpty() && tagNames.isEmpty() -> cardDao.loadTopOtherDueReviewCardByEffortLevels(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis,
-            effortLevels = effortLevels
-        )
-
-        effortLevels.isEmpty() -> cardDao.loadTopOtherDueReviewCardByAnyTags(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis,
-            tagNames = tagNames
-        )
-
-        else -> cardDao.loadTopOtherDueReviewCardByEffortLevelsAndAnyTags(
-            workspaceId = workspaceId,
-            cutoffMillis = cutoffMillis,
-            nowMillis = nowMillis,
-            effortLevels = effortLevels,
-            tagNames = tagNames
-        )
-    }
-}
-
-private suspend fun loadTopNewReviewCardForPredicate(
-    cardDao: CardDao,
-    workspaceId: String,
-    effortLevels: List<EffortLevel>,
-    tagNames: List<String>
-): CardEntity? {
-    return when {
-        effortLevels.isEmpty() && tagNames.isEmpty() -> cardDao.loadTopNewReviewCard(
-            workspaceId = workspaceId
-        )
-
-        effortLevels.isNotEmpty() && tagNames.isEmpty() -> cardDao.loadTopNewReviewCardByEffortLevels(
-            workspaceId = workspaceId,
-            effortLevels = effortLevels
-        )
-
-        effortLevels.isEmpty() -> cardDao.loadTopNewReviewCardByAnyTags(
-            workspaceId = workspaceId,
-            tagNames = tagNames
-        )
-
-        else -> cardDao.loadTopNewReviewCardByEffortLevelsAndAnyTags(
-            workspaceId = workspaceId,
-            effortLevels = effortLevels,
-            tagNames = tagNames
-        )
-    }
-}
-
-@Dao
-interface TagDao {
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insertTags(tags: List<TagEntity>)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertCardTags(cardTags: List<CardTagEntity>)
-
-    @Query("DELETE FROM card_tags WHERE cardId = :cardId")
-    suspend fun deleteCardTags(cardId: String)
-
-    @Query("SELECT * FROM tags WHERE workspaceId = :workspaceId AND name IN (:names)")
-    suspend fun loadTagsByNames(workspaceId: String, names: List<String>): List<TagEntity>
-
-    @Query("SELECT * FROM tags WHERE workspaceId = :workspaceId")
-    suspend fun loadTagsForWorkspace(workspaceId: String): List<TagEntity>
-
-    @Query("SELECT * FROM tags WHERE workspaceId = :workspaceId ORDER BY name ASC, tagId ASC")
-    suspend fun loadTags(workspaceId: String): List<TagEntity>
-
-    @Query("SELECT * FROM tags WHERE workspaceId = :workspaceId ORDER BY name ASC, tagId ASC")
-    fun observeTags(workspaceId: String): Flow<List<TagEntity>>
-
-    @Query(
-        """
-        SELECT DISTINCT tags.name
-        FROM tags
-        INNER JOIN card_tags ON card_tags.tagId = tags.tagId
-        INNER JOIN cards ON cards.cardId = card_tags.cardId
-        WHERE cards.workspaceId = :workspaceId
-            AND tags.workspaceId = cards.workspaceId
-            AND cards.deletedAtMillis IS NULL
-        ORDER BY tags.name ASC
-        """
-    )
-    suspend fun loadReviewTagNames(workspaceId: String): List<String>
-
-    @Query(
-        """
-        SELECT DISTINCT tags.name
-        FROM tags
-        INNER JOIN card_tags ON card_tags.tagId = tags.tagId
-        INNER JOIN cards ON cards.cardId = card_tags.cardId
-        WHERE cards.workspaceId = :workspaceId
-            AND tags.workspaceId = cards.workspaceId
-            AND cards.deletedAtMillis IS NULL
-        ORDER BY tags.name ASC
-        """
-    )
-    fun observeReviewTagNames(workspaceId: String): Flow<List<String>>
-
-    @Query(
-        """
-        SELECT card_tags.*
-        FROM card_tags
-        INNER JOIN cards ON cards.cardId = card_tags.cardId
-        INNER JOIN tags ON tags.tagId = card_tags.tagId
-        WHERE cards.workspaceId = :workspaceId
-            AND tags.workspaceId = :workspaceId
-        ORDER BY cards.createdAtMillis ASC, cards.cardId ASC, card_tags.tagId ASC
-        """
-    )
-    suspend fun loadCardTags(workspaceId: String): List<CardTagEntity>
-
-    @Query("UPDATE card_tags SET cardId = :newCardId WHERE cardId = :oldCardId")
-    suspend fun reassignCardTagsToCard(oldCardId: String, newCardId: String)
-
-    @Query("SELECT EXISTS(SELECT 1 FROM tags WHERE workspaceId = :workspaceId AND LOWER(name) = LOWER(:tagName) LIMIT 1)")
-    suspend fun hasTag(workspaceId: String, tagName: String): Boolean
-
-    @Query("DELETE FROM tags WHERE workspaceId = :workspaceId AND tagId NOT IN (SELECT DISTINCT tagId FROM card_tags)")
-    suspend fun deleteUnusedTags(workspaceId: String)
-
-    @Query("SELECT COUNT(*) FROM tags")
-    suspend fun countTags(): Int
-
-    @Query("DELETE FROM tags")
-    suspend fun deleteAllTags()
-
-    @Query("DELETE FROM card_tags")
-    suspend fun deleteAllCardTags()
-
-    @Query("UPDATE tags SET workspaceId = :newWorkspaceId WHERE workspaceId = :oldWorkspaceId")
-    suspend fun reassignWorkspace(oldWorkspaceId: String, newWorkspaceId: String)
-}
-
-@Dao
-interface ReviewLogDao {
-    @Insert(onConflict = OnConflictStrategy.ABORT)
-    suspend fun insertReviewLog(reviewLog: ReviewLogEntity)
-
-    @Query("SELECT * FROM review_logs")
-    fun observeReviewLogs(): Flow<List<ReviewLogEntity>>
-
-    @Query("SELECT COUNT(*) FROM review_logs")
-    suspend fun countReviewLogs(): Int
-
-    @Query("SELECT COUNT(*) FROM review_logs WHERE workspaceId = :workspaceId")
-    suspend fun countReviewLogs(workspaceId: String): Int
-
-    @Query(
-        """
-        SELECT EXISTS(
-            SELECT 1
-            FROM review_logs
-            WHERE reviewedAtMillis >= :startMillis
-                AND reviewedAtMillis < :endMillis
-            LIMIT 1
-        )
-        """
-    )
-    suspend fun hasReviewLogsBetween(startMillis: Long, endMillis: Long): Boolean
-
-    @Query("SELECT * FROM review_logs ORDER BY reviewedAtMillis DESC")
-    suspend fun loadReviewLogs(): List<ReviewLogEntity>
-
-    @Query("SELECT * FROM review_logs WHERE reviewLogId IN (:reviewLogIds)")
-    suspend fun loadReviewLogs(reviewLogIds: List<String>): List<ReviewLogEntity>
-
-    @Query("SELECT * FROM review_logs WHERE reviewLogId = :reviewLogId LIMIT 1")
-    suspend fun loadReviewLog(reviewLogId: String): ReviewLogEntity?
-
-    @Query("SELECT * FROM review_logs WHERE workspaceId = :workspaceId ORDER BY reviewedAtMillis DESC")
-    suspend fun loadReviewLogs(workspaceId: String): List<ReviewLogEntity>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertReviewLogs(reviewLogs: List<ReviewLogEntity>)
-
-    @Query("UPDATE review_logs SET cardId = :newCardId WHERE workspaceId = :workspaceId AND cardId = :oldCardId")
-    suspend fun reassignReviewLogsToCard(workspaceId: String, oldCardId: String, newCardId: String)
-
-    @Query("DELETE FROM review_logs WHERE reviewLogId IN (:reviewLogIds)")
-    suspend fun deleteReviewLogs(reviewLogIds: List<String>)
-
-    @Query("DELETE FROM review_logs")
-    suspend fun deleteAllReviewLogs()
-
-    @Query("UPDATE review_logs SET workspaceId = :newWorkspaceId WHERE workspaceId = :oldWorkspaceId")
-    suspend fun reassignWorkspace(oldWorkspaceId: String, newWorkspaceId: String)
-}
-
-@Dao
-interface OutboxDao {
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertOutboxEntries(entries: List<OutboxEntryEntity>)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertOutboxEntry(entry: OutboxEntryEntity)
-
-    @Query("SELECT COUNT(*) FROM outbox_entries")
-    fun observeOutboxEntriesCount(): Flow<Int>
-
-    @Query("SELECT COUNT(*) FROM outbox_entries")
-    suspend fun countOutboxEntries(): Int
-
-    @Query("SELECT COUNT(*) FROM outbox_entries WHERE workspaceId = :workspaceId")
-    suspend fun countOutboxEntriesForWorkspace(workspaceId: String): Int
-
-    @Query("SELECT * FROM outbox_entries WHERE workspaceId = :workspaceId ORDER BY createdAtMillis ASC, rowid ASC LIMIT :limit")
-    suspend fun loadOutboxEntries(workspaceId: String, limit: Int): List<OutboxEntryEntity>
-
-    @Query("SELECT * FROM outbox_entries WHERE workspaceId = :workspaceId ORDER BY createdAtMillis ASC, rowid ASC")
-    suspend fun loadAllOutboxEntries(workspaceId: String): List<OutboxEntryEntity>
-
-    @Query(
-        """
-        SELECT * FROM outbox_entries
-        WHERE workspaceId = :workspaceId
-            AND entityType = 'review_event'
-            AND operationType = 'append'
-        ORDER BY createdAtMillis ASC, rowid ASC
-        """
-    )
-    suspend fun loadPendingReviewEventOutboxEntries(workspaceId: String): List<OutboxEntryEntity>
-
-    @Query(
-        """
-        SELECT * FROM outbox_entries
-        WHERE entityType = 'review_event'
-            AND operationType = 'append'
-        ORDER BY createdAtMillis ASC, rowid ASC
-        """
-    )
-    fun observePendingReviewEventOutboxEntries(): Flow<List<OutboxEntryEntity>>
-
-    @Query(
-        """
-        SELECT * FROM outbox_entries
-        WHERE entityType = 'card'
-            AND operationType = 'upsert'
-            AND affectsReviewSchedule = 1
-        ORDER BY workspaceId ASC, createdAtMillis ASC, rowid ASC
-        """
-    )
-    fun observePendingReviewScheduleCardUpsertOutboxEntries(): Flow<List<OutboxEntryEntity>>
-
-    @Query("DELETE FROM outbox_entries WHERE workspaceId = :workspaceId")
-    suspend fun deleteOutboxEntriesForWorkspace(workspaceId: String)
-
-    @Query("DELETE FROM outbox_entries WHERE outboxEntryId IN (:operationIds)")
-    suspend fun deleteOutboxEntries(operationIds: List<String>)
-
-    @Query(
-        """
-        UPDATE outbox_entries
-        SET attemptCount = attemptCount + 1, lastError = :errorMessage
-        WHERE outboxEntryId IN (:operationIds)
-        """
-    )
-    suspend fun markOutboxEntriesFailed(operationIds: List<String>, errorMessage: String)
-
-}
-
-@Dao
-interface SyncStateDao {
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertSyncState(syncState: SyncStateEntity)
-
-    @Query("SELECT * FROM sync_state WHERE workspaceId = :workspaceId LIMIT 1")
-    fun observeSyncState(workspaceId: String): Flow<SyncStateEntity?>
-
-    @Query("SELECT * FROM sync_state WHERE workspaceId = :workspaceId LIMIT 1")
-    suspend fun loadSyncState(workspaceId: String): SyncStateEntity?
-
-    @Query("SELECT * FROM sync_state ORDER BY workspaceId ASC")
-    fun observeSyncStates(): Flow<List<SyncStateEntity>>
-
-    @Query("DELETE FROM sync_state")
-    suspend fun deleteAllSyncState()
-
-    @Query("DELETE FROM sync_state WHERE workspaceId = :workspaceId")
-    suspend fun deleteSyncState(workspaceId: String)
-
-    @Query("UPDATE sync_state SET workspaceId = :newWorkspaceId WHERE workspaceId = :oldWorkspaceId")
-    suspend fun reassignWorkspace(oldWorkspaceId: String, newWorkspaceId: String)
-
-    @Query(
-        """
-        UPDATE sync_state
-        SET lastSyncError = NULL, blockedInstallationId = NULL
-        WHERE blockedInstallationId IS NOT NULL
-        """
-    )
-    suspend fun clearBlockedSyncState()
-}
-
-@Dao
-interface ProgressRemoteCacheDao {
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressSummaryCache(entry: ProgressSummaryCacheEntity)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressSeriesCache(entry: ProgressSeriesCacheEntity)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressReviewScheduleCache(entry: ProgressReviewScheduleCacheEntity)
-
-    @Query("SELECT * FROM progress_summary_cache ORDER BY updatedAtMillis DESC, scopeKey DESC")
-    fun observeProgressSummaryCaches(): Flow<List<ProgressSummaryCacheEntity>>
-
-    @Query("SELECT * FROM progress_series_cache ORDER BY updatedAtMillis DESC, scopeKey DESC")
-    fun observeProgressSeriesCaches(): Flow<List<ProgressSeriesCacheEntity>>
-
-    @Query("SELECT * FROM progress_review_schedule_cache ORDER BY updatedAtMillis DESC, scopeKey DESC")
-    fun observeProgressReviewScheduleCaches(): Flow<List<ProgressReviewScheduleCacheEntity>>
-}
-
-@Dao
-interface ProgressLocalCacheDao {
-    @Query("SELECT * FROM progress_local_day_counts ORDER BY timeZone ASC, workspaceId ASC, localDate ASC")
-    fun observeProgressLocalDayCounts(): Flow<List<ProgressLocalDayCountEntity>>
-
-    @Query(
-        """
-        SELECT * FROM progress_local_day_counts
-        WHERE timeZone = :timeZone AND workspaceId = :workspaceId AND localDate = :localDate
-        LIMIT 1
-        """
-    )
-    suspend fun loadProgressLocalDayCount(
-        timeZone: String,
-        workspaceId: String,
-        localDate: String
-    ): ProgressLocalDayCountEntity?
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressLocalDayCount(entry: ProgressLocalDayCountEntity)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressLocalDayCounts(entries: List<ProgressLocalDayCountEntity>)
-
-    @Query(
-        """
-        DELETE FROM progress_local_day_counts
-        WHERE timeZone = :timeZone AND workspaceId = :workspaceId
-        """
-    )
-    suspend fun deleteProgressLocalDayCounts(
-        timeZone: String,
-        workspaceId: String
-    )
-
-    @Query("DELETE FROM progress_local_day_counts WHERE timeZone = :timeZone")
-    suspend fun deleteProgressLocalDayCounts(timeZone: String)
-
-    @Query("DELETE FROM progress_local_day_counts")
-    suspend fun deleteAllProgressLocalDayCounts()
-
-    @Query(
-        """
-        UPDATE progress_local_day_counts
-        SET workspaceId = :newWorkspaceId
-        WHERE workspaceId = :oldWorkspaceId
-        """
-    )
-    suspend fun reassignWorkspaceProgressLocalDayCounts(
-        oldWorkspaceId: String,
-        newWorkspaceId: String
-    )
-
-    @Query("SELECT * FROM progress_review_history_state ORDER BY workspaceId ASC")
-    fun observeProgressReviewHistoryStates(): Flow<List<ProgressReviewHistoryStateEntity>>
-
-    @Query("SELECT * FROM progress_review_history_state ORDER BY workspaceId ASC")
-    suspend fun loadProgressReviewHistoryStates(): List<ProgressReviewHistoryStateEntity>
-
-    @Query("SELECT * FROM progress_review_history_state WHERE workspaceId = :workspaceId LIMIT 1")
-    suspend fun loadProgressReviewHistoryState(workspaceId: String): ProgressReviewHistoryStateEntity?
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressReviewHistoryState(entry: ProgressReviewHistoryStateEntity)
-
-    @Query("DELETE FROM progress_review_history_state WHERE workspaceId = :workspaceId")
-    suspend fun deleteProgressReviewHistoryState(workspaceId: String)
-
-    @Query("DELETE FROM progress_review_history_state")
-    suspend fun deleteAllProgressReviewHistoryStates()
-
-    @Query(
-        """
-        UPDATE progress_review_history_state
-        SET workspaceId = :newWorkspaceId
-        WHERE workspaceId = :oldWorkspaceId
-        """
-    )
-    suspend fun reassignProgressReviewHistoryState(
-        oldWorkspaceId: String,
-        newWorkspaceId: String
-    )
-
-    @Query("SELECT * FROM progress_local_cache_state ORDER BY timeZone ASC, workspaceId ASC")
-    fun observeProgressLocalCacheStates(): Flow<List<ProgressLocalCacheStateEntity>>
-
-    @Query(
-        """
-        SELECT * FROM progress_local_cache_state
-        WHERE timeZone = :timeZone AND workspaceId = :workspaceId
-        LIMIT 1
-        """
-    )
-    suspend fun loadProgressLocalCacheState(
-        timeZone: String,
-        workspaceId: String
-    ): ProgressLocalCacheStateEntity?
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressLocalCacheState(entry: ProgressLocalCacheStateEntity)
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertProgressLocalCacheStates(entries: List<ProgressLocalCacheStateEntity>)
-
-    @Query(
-        """
-        DELETE FROM progress_local_cache_state
-        WHERE timeZone = :timeZone AND workspaceId = :workspaceId
-        """
-    )
-    suspend fun deleteProgressLocalCacheState(
-        timeZone: String,
-        workspaceId: String
-    )
-
-    @Query("DELETE FROM progress_local_cache_state WHERE timeZone = :timeZone")
-    suspend fun deleteProgressLocalCacheStates(timeZone: String)
-
-    @Query("DELETE FROM progress_local_cache_state")
-    suspend fun deleteAllProgressLocalCacheStates()
-
-    @Query(
-        """
-        UPDATE progress_local_cache_state
-        SET workspaceId = :newWorkspaceId
-        WHERE workspaceId = :oldWorkspaceId
-        """
-    )
-    suspend fun reassignProgressLocalCacheStates(
-        oldWorkspaceId: String,
-        newWorkspaceId: String
-    )
-}
