+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/database/ProgressCardDao.kt
@@ -1,0 +1,18 @@
+package com.flashcardsopensourceapp.data.local.database
+
+import androidx.room.Dao
+import androidx.room.Query
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface ProgressCardDao {
+    @Query(
+        """
+        SELECT cardId, workspaceId, dueAtMillis
+        FROM cards
+        WHERE deletedAtMillis IS NULL
+        ORDER BY workspaceId ASC, dueAtMillis ASC, cardId ASC
+        """
+    )
+    fun observeProgressReviewScheduleCardDueDates(): Flow<List<ProgressReviewScheduleCardDueEntity>>
+}
